refactor(desktop): harden shared role labels and align test fixtures
