+++ apps/desktop/src/pages/kanban-page.tsx
@@ -9,8 +9,8 @@ import { TaskCreateModal } from "@/components/features/task-create-modal";
 import { TaskDetailsSheet } from "@/components/features/task-details-sheet";
 import { Button } from "@/components/ui/button";
 import { useAgentState, useTasksState, useWorkspaceState } from "@/state";
+import { AGENT_ROLE_LABELS } from "@/types";
 import type { AgentSessionState } from "@/types/agent-orchestrator";
-import { AGENT_ROLE_LABELS } from "@/types/agent-role-labels";
 import { firstScenario, kickoffPromptForScenario, SCENARIO_LABELS } from "./agents-page-constants";
 import { useAgentStudioRepoSettings } from "./use-agent-studio-repo-settings";
 import { useSessionStartModalState } from "./use-session-start-modal-state";
