+++ apps/desktop/src/types/agent-role-labels.ts
@@ -1,8 +1,8 @@
 import type { AgentRole } from "@openducktor/core";
 
-export const AGENT_ROLE_LABELS = {
+export const AGENT_ROLE_LABELS: Readonly<Record<AgentRole, string>> = Object.freeze({
   spec: "Spec",
   planner: "Planner",
   build: "Builder",
   qa: "QA",
-} satisfies Record<AgentRole, string>;
+} satisfies Record<AgentRole, string>);
