+++ apps/desktop/src/types/index.ts
@@ -1,3 +1,4 @@
+export { AGENT_ROLE_LABELS } from "./agent-role-labels";
 export type { AgentWorkflowStepState } from "./agent-workflow";
 export type {
   ChecksStateContextValue,
